added token to mutation

+++ src/resolvers/Mutation.js
@@ -53,17 +53,36 @@ const Mutation = {
 
     }}, info);
   },
-  async updateUser(parent, args, { prisma }, info) {
+  async updateUser(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
     return prisma.mutation.updateUser({ data: args.data, where: {
-      id: args.id
+      id: userId
     }}, info)
   },
-  async updatePost(parent, args, { prisma }, info) {
+  async updatePost(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+    const postExists = await prisma.exists.Post({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    })
+    if(!postExists){
+      throw new Error('No post found')
+    }
     return prisma.mutation.updatePost({ data: args.data, where: {
       id: args.id
     }}, info)
   },
-  async updateComment(parent, args, { prisma }, info) {
+  async updateComment(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+    const commentExists = await prisma.exists.Comment({
+      id: args.id,
+      author: userId
+    })
+    if(!commentExists) {
+      throw new Error('Comment does not exists')
+    }
     return prisma.mutation.updateComment({
       data: args.data,
       where: {
@@ -71,13 +90,14 @@ const Mutation = {
       }
     }, info)
   },
-  async createComment(parent, args, { prisma }, info) {
+  async createComment(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
     return prisma.mutation.createComment({
       data: {
         text: args.data.text,
         author: {
           connect: {
-            id: args.data.author
+            id: userId
           }
         },
         post: {
@@ -88,23 +108,39 @@ const Mutation = {
       }
     }, info)
   },
-  async deleteUser(parent, args, { prisma }, info) {
-    const userExists = await prisma.exists.User({ id: args.id})
-    if(!userExists){
-      throw new Error('User not found');
-    }
+  async deleteUser(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+    
     return prisma.mutation.deleteUser({
       where: {
-        id: args.id
+        id: userId
       }
     }, info)
   },
-  async deletePost(parent, args, { prisma }, info) {
+  async deletePost(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+    const postExists = await prisma.exists.Post({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    })
+    if(!postExists){
+      throw new Error('No post found')
+    }
     return prisma.mutation.deletePost({ where: {
       id: args.id }
     }, info)
   },
-  deleteComment(parent, args, { prisma }, info) {
+  async deleteComment(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+    const commentExists = await prisma.exists.Comment({
+      id: args.id,
+      author: userId
+    })
+    if(!commentExists) {
+      throw new Error('Comment does not exists')
+    }
     return prisma.mutation.deleteComment({ where: {
        id: args.id }}, info)
   }
